pyproject.toml
[project]
name = "affine"
version = "1.0.0"
description = "Affine bittensor subnet evaluation rewrite"
readme = "README.md"
authors = ["Affine Contributors <affine@affine.io>"]
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.10",
    "gymnasium>=0.29",
    "numpy>=1.24",
    "typer>=0.12",
    "bittensor>=8.0.0",
    "requests>=2.31",
    "pynacl>=1.5",
    "blake3>=0.4",
]

[project.scripts]
affine = "affine.cli:app"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["affine"]
